+++ zest/releaser/addchangelogentry.py
@@ -8,15 +8,13 @@ import sys
 
 from zest.releaser import baserelease
 from zest.releaser import utils
-from zest.releaser.utils import read_text_file
-from zest.releaser.utils import write_text_file
-
 
 logger = logging.getLogger(__name__)
 
 TODAY = datetime.datetime.today().strftime('%Y-%m-%d')
+HISTORY_HEADER = '%(new_version)s (unreleased)'
 NOTHING_CHANGED_YET = '- Nothing changed yet.'
-COMMIT_MSG = 'Back to development: %(new_version)s'
+COMMIT_MSG = ''
 DEV_VERSION_TEMPLATE = '%(new_version)s.dev0'
 
 DATA = {
@@ -27,7 +25,22 @@ DATA = {
     'reporoot': 'Root of the version control repository',
     'name': 'Name of the project being released',
     'nothing_changed_yet': 'First line in new changelog section',
+    'new_version': 'New development version (so 1.1)',
+    'dev_version': 'New development version with dev marker (so 1.1.dev0)',
+    'commit_msg': (
+        'Message template used when committing. '
+        'Default: same as the message passed on the command line.'),
+    'headings': 'Extracted headings from the history file',
+    'history_file': 'Filename of history/changelog file (when found)',
+    'history_last_release': (
+        'Full text of all history entries of the current release'),
     'history_header': 'Header template used for 1st history header',
+    'history_lines': 'List with all history file lines (when found)',
+    'history_encoding': 'The detected encoding of the history file',
+    'history_insert_line_here': (
+        'Line number where an extra changelog entry can be inserted.'),
+    'dev_version_template': 'Template for dev version number',
+    'message': 'The message we want to add',
 }
 
 
@@ -44,59 +57,33 @@ class AddChangelogEntry(baserelease.Basereleaser):
         self.data.update(dict(
             nothing_changed_yet=NOTHING_CHANGED_YET,
             commit_msg=COMMIT_MSG,
+            history_header=HISTORY_HEADER,
+            message=message.strip(),
             dev_version_template=DEV_VERSION_TEMPLATE))
-        # The message we want to add.
-        self.message = message
 
     def prepare(self):
         """Prepare self.data by asking about new dev version"""
         if not utils.sanity_check(self.vcs):
             logger.critical("Sanity check failed.")
             sys.exit(1)
-        self._grab_version()
-        self._update_history(self.message)
+        self._grab_history()
+        self._get_message()
 
     def execute(self):
         """Make the changes and offer a commit"""
-        self._update_history()
+        self._insert_changelog_entry(self.data['message'])
+        self._write_history()
         self._diff_and_commit()
-        self._push()
 
-    def _update_history(self, message=''):
-        """Update the history file"""
-        version = self.data['new_version']
-        history = self.vcs.history_file()
-        if not history:
-            logger.warn("No history file found")
-            return
-        history_lines, history_encoding = read_text_file(history)
-        history_lines = history_lines.split('\n')
-        headings = utils.extract_headings_from_history(history_lines)
-        if not len(headings):
-            logger.warn("No detectable existing version headings in the "
-                        "history file.")
-            inject_location = 0
-            underline_char = '-'
-        else:
-            first = headings[0]
-            inject_location = first['line']
-            underline_line = first['line'] + 1
-            try:
-                underline_char = history_lines[underline_line][0]
-            except IndexError:
-                logger.debug("No character on line below header.")
-                underline_char = '-'
-        header = '%s (unreleased)' % version
-        inject = [header,
-                  underline_char * len(header),
-                  '',
-                  self.data['nothing_changed_yet'],
-                  '',
-                  '']
-        history_lines[inject_location:inject_location] = inject
-        contents = '\n'.join(history_lines)
-        write_text_file(history, contents, history_encoding)
-        logger.info("Injected new section into the history: %r", header)
+    def _get_message(self):
+        """Get changelog message and commit message."""
+        message = self.data['message']
+        while not message:
+            q = "What is the changelog message? "
+            message = utils.get_input(q)
+        self.data['message'] = message
+        if not self.data['commit_msg']:
+            self.data['commit_msg'] = message
 
 
 def datacheck(data):
@@ -111,5 +98,5 @@ def main():
         help="Text of changelog entry")
     options = utils.parse_options(parser)
     utils.configure_logging()
-    addchangelogentry = AddChangelogEntry()
+    addchangelogentry = AddChangelogEntry(message=options.message)
     addchangelogentry.run()